apps/workers/requirements.txt
celery==5.3.4
redis==5.0.1
nats-py==2.3.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
crewai==0.1.0
langchain==0.0.350
openai==1.3.7
anthropic==0.7.7
httpx==0.25.2
python-dotenv==1.0.0
structlog==23.2.0
prometheus-client==0.19.0
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
